pyproject.toml
[project]
name = "vitasys"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aioredis==1.3.1",
    "amqp==5.3.1",
    "ansicon==1.89.0",
    "argon2-cffi==25.1.0",
    "argon2-cffi-bindings==25.1.0",
    "arrow==1.4.0",
    "asgiref==3.9.2",
    "async-timeout==5.0.1",
    "attrs==25.4.0",
    "autobahn==25.10.2",
    "automat==25.4.16",
    "billiard==4.2.2",
    "blessed==1.23.0",
    "celery==5.5.3",
    "certifi==2025.10.5",
    "cffi==2.0.0",
    "channels==3.0.5",
    "channels-redis==3.3.1",
    "charset-normalizer==3.4.4",
    "click==8.3.0",
    "click-didyoumean==0.3.1",
    "click-plugins==1.1.1.2",
    "click-repl==0.3.0",
    "colorama==0.4.6",
    "constantly==23.10.4",
    "cron-descriptor==2.0.6",
    "cryptography==46.0.3",
    "daphne==3.0.2",
    "distlib==0.4.0",
    "dj-database-url==3.0.1",
    "django==5.2.7",
    "django-celery-beat==2.8.1",
    "django-celery-results==2.6.0",
    "django-cors-headers==4.9.0",
    "django-extensions==4.1",
    "django-jazzmin==3.0.1",
    "django-picklefield==3.3",
    "django-q2==1.8.0",
    "django-redis==6.0.0",
    "django-timezone-field==7.1",
    "djangorestframework==3.16.1",
    "djangorestframework-simplejwt==5.5.1",
    "dotenv==0.9.9",
    "drf-yasg==1.21.11",
    "filelock==3.19.1",
    "hiredis==3.3.0",
    "hyperlink==21.0.0",
    "idna==3.11",
    "incremental==24.7.2",
    "inflection==0.5.1",
    "jinxed==1.3.0",
    "kombu==5.5.4",
    "msgpack==1.1.2",
    "oauthlib==3.3.1",
    "packaging==25.0",
    "pillow==12.0.0",
    "platformdirs==4.4.0",
    "prompt-toolkit==3.0.52",
    "psycopg2-binary==2.9.11",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.23",
    "pyjwt==2.10.1",
    "pyopenssl==25.3.0",
    "python-crontab==3.3.0",
    "python-dateutil==2.9.0.post0",
    "python-decouple==3.8",
    "python-dotenv==1.1.1",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "redis==7.0.1",
    "requests==2.32.5",
    "requests-oauthlib==2.0.0",
    "service-identity==24.2.0",
    "setuptools==80.9.0",
    "six==1.17.0",
    "sqlparse==0.5.3",
    "twisted==25.5.0",
    "txaio==25.9.2",
    "typing-extensions==4.15.0",
    "tzdata==2025.2",
    "uritemplate==4.2.0",
    "urllib3==2.5.0",
    "vine==5.1.0",
    "virtualenv==20.34.0",
    "wcwidth==0.2.14",
    "whitenoise==6.11.0",
    "zope-interface==8.0.1",
]
